pyproject.toml
[project]
name = "nekobot"
version = "1.0.0"
description = "A modular chatbot framework for NapCat"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "websockets>=12.0",
    "aiohttp>=3.9.0",
    "pydantic>=2.5.0",
    "pyyaml>=6.0.1",
    "loguru>=0.7.2",
    "certifi>=2025.11.12",
    "quart>=0.20.0",
    "quart-cors>=0.7.0",
    "python-jose>=3.5.0",
    "passlib[bcrypt]>=1.7.4",
    "tomli>=0.2.1",
    "psutil>=6.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.ruff]
line-length = 88
target-version = "py310"
exclude = [".venv", "venv", "env", ".git", "__pycache__", "build", "dist"]

[tool.ruff.lint]
ignore = ["F403", "F405", "E501", "ASYNC230"]
select = ["F", "E", "W", "ASYNC", "C4", "Q"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.14.10",
]
